feat(v0.0.2): add actor crud

+++ crud-spring/src/main/java/com/example/crud_spring/services/ActorService.java
@@ -1,5 +1,7 @@
 package com.example.crud_spring.services;
 
+import com.example.crud_spring.helpers.ApiRestFullResponses;
+import static com.example.crud_spring.helpers.ApiRestFullResponses.*;
 import com.example.crud_spring.models.Actor;
 import com.example.crud_spring.repositories.IActorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,42 +9,73 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
+/**
+ * Create CRUD whit Spring
+ *
+ * @author Alex T.H.
+ * @version v0.0.1
+ * @since 21.0.0 2024-08-23
+ */
 @Service
 public class ActorService {
 
     @Autowired
     private IActorRepository actorRepository;
 
-    public Actor getActorById(Integer id) {
-        return actorRepository.findById(id).get();
+    public ApiRestFullResponses getActorById(Integer id) {
+        Optional<Actor> actor = actorRepository.findById(id);
+        return actor.map(value -> response("OK", 200, value))
+                .orElseGet(() -> response("ERROR", 404, "ACTOR_NOT_FOUND"));
     }
 
-    public List<Actor> getAllActors(){
-        return actorRepository.findAll();
+    public ApiRestFullResponses getAllActors(){
+        List<Actor> actors = actorRepository.findAll();
+        if(actors.isEmpty()) return response("NULL", 404, "ACTORS_NOT_FOUND");
+        return response("OK", 200, actors);
     }
 
-    public Actor save(Actor request){
+    public ApiRestFullResponses save(Actor request){
         Actor actor = new Actor();
         actor.setActorId(request.getActorId());
         actor.setFirstName(request.getFirstName());
         actor.setLastName(request.getLastName());
         actor.setLastUpdate(LocalDateTime.now().toString());
-        return actorRepository.save(actor);
+        actorRepository.save(actor);
+        return response("OK", 200, actor);
     }
 
-    public Actor put(Actor actor, Integer id ){
-        Actor currentActor = actorRepository.findById(id).get();
-        currentActor.setFirstName(actor.getFirstName());
-        currentActor.setLastName(actor.getLastName());
-        currentActor.setLastUpdate(actor.getLastUpdate());
-        return actorRepository.save(currentActor);
+    public ApiRestFullResponses put(Actor request, Integer id ){
+        Optional<Actor> actorOptional = actorRepository.findById(id);
+
+        if(actorOptional.isEmpty()) return response("ERROR", 404, "ACTOR_NOT_FOUND");
+
+        Actor actor = actorOptional.get();
+
+        if(request.getFirstName() != null){
+            actor.setFirstName(request.getFirstName());
+        }
+
+        if(request.getLastName() != null){
+            actor.setLastName(request.getLastName());
+        }
+
+        if(request.getLastUpdate() != null){
+            actor.setLastUpdate(request.getLastUpdate());
+        }
+
+        actorRepository.save(actor);
+
+        return response("OK", 200, actor);
     }
 
-    public String remove(Integer id) {
-        Actor actor = actorRepository.findById(id).get();
+    public ApiRestFullResponses remove(Integer id) {
+        Optional<Actor> actorOptional = actorRepository.findById(id);
+        if(actorOptional.isEmpty()) return response("ERROR", 404, "ACTOR_NOT_FOUND");
+        Actor actor = actorOptional.get();
         actorRepository.delete(actor);
-        return "ACTOR HAS DELETED SUCCESSFULLY";
+        return response("OK", 200, "ACTOR_HAS_DELETED_SUCCESSFULLY");
     }
